Rename testing file
Added testing for BeyondDFTMethod
